uporządkowanie plików pod panel administratora
