+++ main.py
@@ -7,6 +7,7 @@ from flask import Flask
 
 # importy nasze
 from views import pages, login_manager
+from views_admin import pages_admin
 from mail_sending import mail
 from models import db
 from config import DB, APP, EMAIL
@@ -24,6 +25,7 @@ app.config['MAIL_PASSWORD'] = EMAIL.PASSWORD
 app.config['MAIL_DEFAULT_SENDER'] = EMAIL.DEFAULT_SENDER
 
 app.register_blueprint(pages)
+app.register_blueprint(pages_admin)
 login_manager.init_app(app)
 mail.init_app(app)
 db.init_app(app)
