+++ templates/navbar.html
@@ -32,6 +32,6 @@
     {% endif %}
     <div style="clear: both"></div>
     {% if current_user.admin %}
-        <a href="{{ url_for('pages.admin') }}">panel administratora</a>
+        <a href="{{ url_for('pages_admin.admin') }}">panel administratora</a>
     {% endif %}
 </div>
